hsa/radeon: Fix bug which prevented suspend during wait-on-events
This patch fixes a bug which prevented the OS to enter suspend mode in case a
user process was waiting on an event (using ioctl KFD_IOC_WAIT_EVENTS).

The fix checks if a signal was sent to the waiting process and if so, terminates
the wait and sends EAGAIN to the calling process, allowing it to return to
userspace and do the suspend.

Signed-off-by: Oded Gabbay <oded.gabbay@amd.com>