amdkfd: Fixed calculation of gart buffer size
This patch enhances the calculation of the gart buffer size that is needed for
the amdkfd operation.

It takes into account maximum number of MQDs, runlist packets, kernel queues
and reserves 512KB for other misc allocations.

The total size is just shy of 4MB, for 32 processes and 128 queues per process.

v2: Calculation of gart buffer size was slightly fixed

Reviewed-by: Alexey Skidanov <Alexey.skidanov@amd.com>
Signed-off-by: Oded Gabbay <oded.gabbay@amd.com>