amd: Add new kfd-->kgd interface for gart usage
This patch adds two new functions to the kfd-->kgd interface:

init_gtt_mem_allocation, which allocate a buffer on the gart aperture.
The buffer can be used for mqds, hpds, kernel queue, fence and runlists.

free_gtt_mem, which frees a buffer that was allocated on the gart aperture.

The init_gtt_mem function is called only once per GPU device. It allocates a
buffer which is large enough for all the kfd needs for that device.

v2: Return the gpu address and cpu ptr as arguments

Reviewed-by: Alexey Skidanov <Alexey.skidanov@amd.com>
Signed-off-by: Oded Gabbay <oded.gabbay@amd.com>